+++ apps/pr-deploy-site/just.config.ts
@@ -1,14 +1,39 @@
 import fs from 'fs';
 import path from 'path';
 import { preset, just } from '@uifabric/build';
-import { findRepoDeps, findGitRoot } from '@uifabric/build/monorepo/index';
+import { findGitRoot, getAllPackageInfo } from '@uifabric/build/monorepo/index';
 
-const { task, copyInstructionsTask, copyInstructions } = just;
+const { series, task, copyInstructionsTask, copyInstructions } = just;
 
 const gitRoot = findGitRoot();
-let instructions = copyInstructions.copyFilesToDestinationDirectory(['index.html', 'chiclet-test.html'], 'dist');
+let instructions = copyInstructions.copyFilesToDestinationDirectory(
+  ['pr-deploy-site.js', 'pr-deploy-site.css', 'chiclet-test.html'],
+  'dist',
+);
 
-const repoDeps = findRepoDeps();
+// If you are adding a new tile into this site, please make sure it is also listed in the siteInfo of
+// `pr-deploy-site.js`
+//
+// Dependencies are listed here and NOT in package.json because we do not want to allow for partial builds for scoping
+const dependencies = [
+  '@fluentui/docs',
+  '@fluentui/react-button',
+  '@fluentui/react-next',
+  '@fluentui/react-tabs',
+  '@uifabric/charting',
+  '@uifabric/date-time',
+  '@uifabric/experiments',
+  '@uifabric/fabric-website',
+  '@uifabric/fabric-website-resources',
+  'office-ui-fabric-react',
+  'perf-test',
+  'theming-designer',
+  'todo-app',
+];
+
+const allPackages = getAllPackageInfo();
+const repoDeps = dependencies.map(dep => allPackages[dep]);
+const deployedPackages = new Set<string>();
 repoDeps.forEach(dep => {
   const distPath = path.join(gitRoot, dep.packagePath, 'dist');
 
@@ -24,12 +49,15 @@ repoDeps.forEach(dep => {
           path.join('dist', path.basename(dep.packagePath)),
         ),
       );
+      deployedPackages.add(dep.packageJson.name);
     } else if (dep.packageJson.name === '@fluentui/docs') {
       instructions.push(...copyInstructions.copyFilesInDirectory(sourcePath, path.join('dist', 'react-northstar')));
+      deployedPackages.add(dep.packageJson.name);
     } else {
       instructions.push(
         ...copyInstructions.copyFilesInDirectory(sourcePath, path.join('dist', path.basename(dep.packagePath))),
       );
+      deployedPackages.add(dep.packageJson.name);
     }
   }
 
@@ -43,9 +71,24 @@ repoDeps.forEach(dep => {
         path.join('dist', path.basename(dep.packagePath)),
       ),
     );
+    deployedPackages.add(dep.packageJson.name);
   }
 });
 
 preset();
 
-task('bundle', copyInstructionsTask({ copyInstructions: instructions }));
+/**
+ * Renders a site with tiles that are potentially from a partial set of deployed packages
+ */
+task('generate:index', () => {
+  const indexContent = fs.readFileSync(path.join(__dirname, './index.html'), 'utf-8');
+  fs.writeFileSync(
+    path.join('dist', 'index.html'),
+    indexContent.replace('/* insert packages here */', JSON.stringify([...deployedPackages])),
+  );
+});
+
+/**
+ * Copies all the built (potentially partially) dist files and then generates a index HTML for it
+ */
+task('generate:site', series(copyInstructionsTask({ copyInstructions: instructions }), 'generate:index'));