+++ packages/fluentui/react-northstar/src/themes/teams/componentVariables.ts
@@ -65,6 +65,11 @@ export { loaderVariables as Loader } from './components/Loader/loaderVariables';
 export { itemLayoutVariables as ItemLayout } from './components/ItemLayout/itemLayoutVariables';
 
 export { listItemVariables as ListItem } from './components/List/listItemVariables';
+export { listItemContentVariables as ListItemContent } from './components/List/listItemContentVariables';
+export { listItemContentMediaVariables as ListItemContentMedia } from './components/List/listItemContentMediaVariables';
+export { listItemHeaderVariables as ListItemHeader } from './components/List/listItemHeaderVariables';
+export { listItemHeaderMediaVariables as ListItemHeaderMedia } from './components/List/listItemHeaderMediaVariables';
+export { listItemMediaVariables as ListItemMedia } from './components/List/listItemMediaVariables';
 
 export { menuVariables as Menu } from './components/Menu/menuVariables';
 export { menuItemVariables as MenuItem } from './components/Menu/menuItemVariables';
@@ -142,6 +147,7 @@ export { cardTopControlsVariables as CardTopControls } from './components/Card/c
 export { cardExpandableBoxVariables as CardExpandableBox } from './components/Card/cardExpandableBoxVariables';
 
 export { datepickerVariables as Datepicker } from './components/Datepicker/datepickerVariables';
+export { datepickerCalendarCellVariables as DatepickerCalendarCell } from './components/Datepicker/datepickerCalendarCellVariables';
 export { datepickerCalendarVariables as DatepickerCalendar } from './components/Datepicker/datepickerCalendarVariables';
 export { datepickerCalendarHeaderVariables as DatepickerCalendarHeader } from './components/Datepicker/datepickerCalendarHeaderVariables';
 export { datepickerCalendarHeaderCellVariables as DatepickerCalendarHeaderCell } from './components/Datepicker/datepickerCalendarHeaderCellVariables';