+++ packages/fluentui/react-northstar/test/specs/components/List/ListItemMedia-test.tsx
@@ -1,0 +1,6 @@
+import { ListItemMedia } from 'src/components/List/ListItemMedia';
+import { isConformant } from 'test/specs/commonTests';
+
+describe('ListItemMedia', () => {
+  isConformant(ListItemMedia, { constructorName: 'ListItemMedia' });
+});